Friday update!
Hi all!

Thank you for being some of the first to contribute to our firmware repository here in this first week.

Special thanks to user marcus-j-davies for creating YAML files for firmware requests and bug reports.

These have been a big help in keeping us organized as we make plans each week to release requested firmware in a way that is information-rich and easy to use.

Here are the changes we're bringing about in this first major update:

1. Release firmware for:
	- Enbrighten-GE
		- ZW1002\14288 - In-Wall Smart Outlet
		- ZW3004\14295 - In-Wall Smart Toggle Dimmer
		- ZW3005\14294 - In-Wall Smart Paddle Dimmer
		- ZW3104\28167 - Plug-in 1-Outlet Smart Dimmer
		- ZW4002\14287 - In-Wall Smart Fan Speed Control
		- ZW4003\14292 - In-Wall Smart Toggle Switch
		- ZW4005\14291 - In-Wall Smart Paddle Switch
		- ZW4103\28169 - Plug-in 1-Outlet Smart Switch
	- Honeywell
		- ZW3005\39351 (OG) - In-Wall Smart Paddle Dimmer
		- ZW4003\39354 (OG) - In-Wall Smart Toggle Switch
		- ZW4005\39348 (OG) - In-Wall Smart Paddle Switch

2. Replace README.txt files for all previous versions with Markdown format (.md). All readme files will be in .md format moving forward.

3. Replace <Model Number>_Revision_History.xlsx files for all previous versions with Comma Separated Values format (.csv). All revision history files will be in .csv format moving forward.