Thursday mini-update
Due to the holiday weekend, we will not be issuing our usual Friday major update this week.

What was originally hoped to be tomorrow's major update should come to be our final batch of firmware uploaded by end-of-week next week. In the meantime, we are discussing what the future of this repository will look like since we have continued to enjoy this community-facing experience.

Enjoy this mini-update, however, if you are a user of:
	- Enbrighten-GE
		- ZW3004\14296 - In-Wall Smart Toggle Dimmer (Lt. Alm)

Other fixes:

2. Documented the existence of pre-6.51.06 SDK-developed firmware releases for some of our Jasco-branded products

3. Updated several readmes and repackaged into new .zip archives